requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
Authlib==1.3.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
exceptiongroup==1.2.2
grpcio==1.67.0
grpcio-health-checking==1.67.0
grpcio-tools==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.0
idna==3.10
protobuf==5.28.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
validators==0.34.0
weaviate-client==4.9.0
